+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.10.1 (2017-04-25)
 
 - Fix bug with variable in loop using starting with the container name (#165)
-
+- Allow whitespace in macros/filters params
 
 ## 0.10.0 (2017-04-21)